removed implicit conversions

+++ implementations/data-structures/SegmentTreeLazy.h
@@ -11,7 +11,7 @@ struct SegmentTree {
 
     SegmentTree(int _n) : n(_n), st(4*n), lazy(4*n) {}
 
-    SegmentTree(const vector<int> &_a) : n(_a.size()), a(_a), st(4*n), lazy(4*n) {
+    SegmentTree(const vector<int> &_a) : n((int) _a.size()), a(_a), st(4*n), lazy(4*n) {
         build(1, 0, n-1);
     }
